perf: reduce load times when sorting and filtering iimi results
